refactor: move manager request types to separate file to avoid antipattern
