+++ python/sglang/srt/disaggregation/prefill.py
@@ -42,6 +42,7 @@ from sglang.srt.disaggregation.utils import (
     poll_and_all_reduce,
     prepare_abort,
 )
+from sglang.srt.managers.request_types import FINISH_LENGTH, RequestStage
 from sglang.srt.mem_cache.common import release_kv_cache
 from sglang.srt.mem_cache.memory_pool import (
     HybridLinearKVPool,
@@ -175,8 +176,6 @@ class PrefillBootstrapQueue:
         return kv_manager
 
     def add(self, req: Req, num_kv_heads: int) -> None:
-        from sglang.srt.managers.schedule_batch import RequestStage
-
         if self._check_if_req_exceed_kv_capacity(req):
             return
 
@@ -229,8 +228,6 @@ class PrefillBootstrapQueue:
         return_failed_reqs: For PP, on rank 0, also return the failed reqs to notify the next rank
         rids_to_check: For PP, on rank > 0, check the rids from the previous rank has consensus with the current rank.
         """
-        from sglang.srt.managers.schedule_batch import RequestStage
-
         bootstrapped_reqs = []
         failed_reqs = []
         indices_to_remove = set()
@@ -391,8 +388,6 @@ class SchedulerDisaggregationPrefillMixin:
         Transfer kv for prefill completed requests and add it into disagg_prefill_inflight_queue
         Adapted from process_batch_result_prefill
         """
-        from sglang.srt.managers.schedule_batch import RequestStage
-
         (
             logits_output,
             next_token_ids,
@@ -510,8 +505,6 @@ class SchedulerDisaggregationPrefillMixin:
         Poll the requests in the middle of transfer. If done, return the request.
         rids_to_check: For PP, on rank > 0, check the rids from the previous rank has consensus with the current rank.
         """
-        from sglang.srt.managers.schedule_batch import FINISH_LENGTH, RequestStage
-
         if len(self.disagg_prefill_inflight_queue) == 0:
             return []
